refactor: add typed hooks, move store to file
