+++ src/components/cv-form/CVForm.tsx
@@ -5,13 +5,8 @@ import Skills from '../skills/Skills';
 import Work from '../work/Work';
 
 function CVForm() {
-  const handleFormChange = (e: unknown) => {
-    console.log('on form change', e);
-  };
-
   return (
     <Form
-      onValuesChange={handleFormChange}
       name='cv-form'
       style={{ marginTop: '20px', paddingBottom: '50px' }}
       initialValues={{ remember: true }}
