+++ src/hooks/personal-info/usePersonalInfoDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from 'react-redux';
+import type { AppDispatch } from '../../store/personal-info/personalInfoStore';
+
+export const usePersonalInfoDispatch: () => AppDispatch = useDispatch;
