+++ src/store/personal-info/personalInfoSlice.ts
@@ -1,4 +1,3 @@
-import { configureStore } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
 
@@ -20,12 +19,5 @@ export const personalInfoSlice = createSlice({
   },
 });
 
-export const store = configureStore({
-  reducer: {
-    personalInfo: personalInfoSlice.reducer,
-  },
-});
-
 export const { setName } = personalInfoSlice.actions;
 export default personalInfoSlice.reducer;
-export type RootState = ReturnType<typeof store.getState>;
